confirmations, erase times, bigger scramble

+++ tabs.go
@@ -1,21 +1,26 @@
 package main
 
 import (
+	"encoding/csv"
 	"fmt"
-	"fyne.io/fyne/v2"
-	"fyne.io/fyne/v2/container"
-	"fyne.io/fyne/v2/data/binding"
-	"fyne.io/fyne/v2/layout"
-	"fyne.io/fyne/v2/widget"
+	"image/color"
 	"os"
 	"slices"
 	"time"
+
+	"fyne.io/fyne/v2"
+	"fyne.io/fyne/v2/canvas"
+	"fyne.io/fyne/v2/container"
+	"fyne.io/fyne/v2/data/binding"
+	"fyne.io/fyne/v2/dialog"
+	"fyne.io/fyne/v2/layout"
+	"fyne.io/fyne/v2/widget"
 )
 
-func setTabs() *container.AppTabs {
+func setTabs(w fyne.Window) *container.AppTabs {
 	tabs := container.NewAppTabs(
 		container.NewTabItem("Timer", container.NewVBox(timerTab())),
-		container.NewTabItem("Solves", container.NewStack(solvesTab())),
+		container.NewTabItem("Solves", container.NewStack(solvesTab(w))),
 		// container.NewTabItem("Stats", container.NewCenter(statsTab())),
 	)
 	return tabs
@@ -24,7 +29,7 @@ func setTabs() *container.AppTabs {
 func timerTab() fyne.CanvasObject {
 	ao5, ao12, ao50, ao100, aoAll := binding.NewString(), binding.NewString(), binding.NewString(), binding.NewString(), binding.NewString()
 	go func() {
-		for {
+		for range time.Tick(time.Millisecond * 500) {
 			newSlice := readFile(0)
 			slices.Reverse(newSlice)
 			ao5.Set(getAverage(5, newSlice))
@@ -32,14 +37,13 @@ func timerTab() fyne.CanvasObject {
 			ao50.Set(getAverage(50, newSlice))
 			ao100.Set(getAverage(100, newSlice))
 			aoAll.Set(getAverage(0, newSlice))
-			time.Sleep(time.Millisecond * 500)
 		}
 	}()
 	time.Sleep(time.Millisecond * 500)
 	return container.New(layout.NewGridLayoutWithRows(3), container.NewCenter(gen_scramble_cont()), container.NewCenter(timer), container.NewGridWithRows(5, gen_avg_cont(5, ao5), gen_avg_cont(12, ao12), gen_avg_cont(50, ao50), gen_avg_cont(100, ao100), gen_avg_cont(0, aoAll)))
 }
 
-func solvesTab() fyne.CanvasObject {
+func solvesTab(window fyne.Window) fyne.CanvasObject {
 	list1 := widget.NewList(
 		func() int {
 			return len(decodeFile())
@@ -61,17 +65,61 @@ func solvesTab() fyne.CanvasObject {
 			o.(*widget.Label).SetText(decodeFile()[i][1])
 		})
 	erase := widget.NewButton("Erase all", func() {
-		os.Truncate(filePath, 0)
-		list1.Refresh()
-		list2.Refresh()
-		fmt.Println("Erased all times")
+		dialog.ShowConfirm("Confirmation", "Are you sure you want to delete all saved times?\nThis cannot be undone!", func(choice bool) {
+			if choice {
+				os.Truncate(filePath, 0)
+				list1.Refresh()
+				list2.Refresh()
+				fmt.Println("Erased all times")
+			}
+		}, window)
 	})
 	list1.OnSelected = func(id widget.ListItemID) {
-		list1.Refresh()
+		dialog.ShowConfirm("Confirmation", "Are you sure you want to delete this time?\nThis cannot be undone!", func(choice bool) {
+			if choice {
+				var newSlice [][]string
+				newSlice = slices.Delete(decodeFile(), id, id+1)
+				if len(newSlice) == 0 {
+					os.Truncate(filePath, 0)
+					list1.Refresh()
+					list2.Refresh()
+					list1.UnselectAll()
+					return
+				}
+				os.Truncate(filePath, 0)
+				f, _ := os.OpenFile(filePath, os.O_WRONLY, 0644)
+				writer := csv.NewWriter(f)
+				writer.WriteAll(newSlice)
+				defer f.Close()
+				list1.Refresh()
+				list2.Refresh()
+				list1.UnselectAll()
+			}
+		}, window)
 		list1.UnselectAll()
 	}
 	list2.OnSelected = func(id widget.ListItemID) {
-		list2.Refresh()
+		dialog.ShowConfirm("Confirmation", "Are you sure you want to delete this time?\nThis cannot be undone!", func(choice bool) {
+			if choice {
+				var newSlice [][]string
+				newSlice = slices.Delete(decodeFile(), id, id+1)
+				if len(newSlice) == 0 {
+					os.Truncate(filePath, 0)
+					list2.Refresh()
+					list1.Refresh()
+					list2.UnselectAll()
+					return
+				}
+				os.Truncate(filePath, 0)
+				f, _ := os.OpenFile(filePath, os.O_WRONLY, 0644)
+				writer := csv.NewWriter(f)
+				writer.WriteAll(newSlice)
+				defer f.Close()
+				list1.Refresh()
+				list2.Refresh()
+				list2.UnselectAll()
+			}
+		}, window)
 		list2.UnselectAll()
 	}
 	erasecont := container.NewHBox(erase)
@@ -121,18 +169,18 @@ func gen_avg_cont(num uint8, data binding.String) fyne.CanvasObject {
 }
 
 func gen_scramble_cont() fyne.CanvasObject {
-	data := binding.NewString()
+	currentScramble = getScramble()
 	go func() {
-		for {
+		for range time.Tick(time.Millisecond * 500) {
 			if timesSaved {
 				scr := getScramble()
-				data.Set(scr)
+				scrambleText.Text = scr
 				currentScramble = scr
 				timesSaved = false
 			}
-			time.Sleep(time.Millisecond * 500)
 		}
 	}()
-	wid := widget.NewLabelWithData(data)
-	return wid
+	scrambleText = canvas.NewText(currentScramble, color.White)
+	scrambleText.TextSize = 20
+	return scrambleText
 }
